+++ packages/core/avatar-group/docs/01-avatar-group.js
@@ -1,6 +1,7 @@
 // @flow
 import React from 'react';
 import { md, Example, Props } from '@atlaskit/docs';
+import { Prop } from 'pretty-proptypes';
 
 export default md`
   \`AvatarGroup\` is a wrapper around avatars designed to render a collection
@@ -34,6 +35,26 @@ ${(
   <Props
     heading="Avatar Group Props"
     props={require('!!extract-react-types-loader!../src/components/AvatarGroup')}
+    overrides={{
+      avatar: props => {
+        // Currently prett-proptypes does not have a good print type for function
+        // calls, so we are overriding how this is printed. AK-5133 should resolve
+        // this.
+        /* eslint-disable */
+        if (
+          props &&
+          props.typeValue &&
+          props.typeValue.typeParams &&
+          props.typeValue.typeParams.params
+        ) {
+          props.typeValue.typeParams.params = [
+            { kind: 'id', name: '@atlaskit/avatar' },
+          ];
+        }
+        /* eslint-enable */
+        return <Prop {...props} defaultValue="@atlaskit/avatar" />;
+      },
+    }}
   />
 )}
 `;