+++ packages/core/navigation-next/src/index.js
@@ -7,9 +7,9 @@ export {
   default as GlobalItemPrimitive,
 } from './components/GlobalItem/primitives';
 export { default as GlobalNav } from './components/GlobalNav';
-export { default as Item } from './components/Item';
 export { default as Group } from './components/Group';
 export { default as GroupHeading } from './components/GroupHeading';
+export { default as Item } from './components/Item';
 export { default as ItemPrimitive } from './components/Item/primitives';
 export { default as ItemAvatar } from './components/ItemAvatar';
 export { default as PeekToggleItem } from './components/PeekToggleItem';
@@ -23,13 +23,17 @@ export { default as Switcher } from './components/Switcher';
 
 /** State */
 export { NavigationProvider } from './provider';
-export { UIState, UIStateSubscriber, withNavigationUI } from './ui-state';
 export {
-  ViewState,
-  ViewStateSubscriber,
-  withNavigationViews,
+  UIController,
+  UIControllerSubscriber,
+  withNavigationUI,
+} from './ui-controller';
+export {
+  ViewController,
+  ViewControllerSubscriber,
+  withNavigationViewController,
   viewReducerUtils,
-} from './view-state';
+} from './view-controller';
 
 /** Renderer */
 export { default as ViewRenderer } from './renderer';
@@ -38,5 +42,5 @@ export { default as ViewRenderer } from './renderer';
 export { dark, light, settings, modeGenerator } from './theme';
 
 /** Types */
-export { GlobalItemProps } from './components/GlobalItem';
+export { GlobalItemProps } from './components/GlobalItem/types';
 export { GlobalTheme } from './theme';