+++ packages/core/popper/src/index.js
@@ -1,0 +1,2 @@
+// @flow
+export { Manager, Reference, Popper } from './Popper';