+++ packages/editor/editor-core/src/plugins/panel/index.tsx
@@ -7,17 +7,18 @@ import { getToolbarConfig } from './toolbar';
 
 const panelPlugin: EditorPlugin = {
   nodes() {
-    return [{ rank: 1100, name: 'panel', node: panel }];
+    return [{ name: 'panel', node: panel }];
   },
 
   pmPlugins() {
-    return [{ rank: 1110, plugin: createPlugin }];
+    return [{ name: 'panel', plugin: createPlugin }];
   },
 
   pluginsOptions: {
     quickInsert: [
       {
         title: 'Panel',
+        priority: 1000,
         icon: () => <InfoIcon label="Panel" />,
         action(insert, state) {
           return insert(