+++ packages/media/media-card/src/index.ts
@@ -17,6 +17,7 @@ import { CardAction } from './actions';
 // the only components we expose to consumers is Card, CardView and CardList
 export * from './root/card';
 export * from './root/cardView';
+export * from './root/domain';
 export * from './list';
 export * from './actions';
 // TODO: don't expose this directly https://jira.atlassian.com/browse/FIL-4396