+++ packages/media/media-ui/src/index.ts
@@ -1,9 +1,8 @@
+import * as BlockCard from './BlockCard';
+import * as InlineCard from './InlineCard';
+
 export { Ellipsify, EllipsifyProps } from './ellipsify';
-export { ErrorCard, ErrorCardProps } from './ErrorCard';
-export { CardFrame, CardFrameProps } from './CardFrame';
-export { CardPreview, CardPreviewProps } from './CardPreview';
-export { LinkIcon, LinkIconProps } from './LinkIcon';
-export { IconImage, IconImageProps } from './IconImage';
-export { newCardDetailsHeight } from './newCardDetailsHeight';
 export { toHumanReadableMediaSize } from './humanReadableSize';
 export * from './mixins';
+
+export { BlockCard, InlineCard };