+++ packages/media/media-ui/src/mixins.ts
@@ -1,22 +1,5 @@
 import { keyframes } from 'styled-components';
-import {
-  akColorN900,
-  akColorN300,
-  akBorderRadius,
-} from '@atlaskit/util-shared-styles';
-
-export const cardTitle = `
-  color: ${akColorN900};
-  font-size: 16px;
-  font-weight: 500;
-  line-height: ${20 / 16};
-`;
-
-export const cardDescription = `
-  color: ${akColorN300};
-  font-size: 12px;
-  line-height: ${16 / 12};
-`;
+import { borderRadius as akBorderRadius } from '@atlaskit/theme';
 
 export const ellipsis = (maxWidth: string | number = '100%') => {
   const unit = typeof maxWidth === 'number' ? 'px' : '';