+++ packages/media/smart-card/__tests__/extractBlockPropsFromJSONLD/extractBlockViewPropsFromSpreadsheet.test.ts
@@ -1,0 +1,11 @@
+import { spreadsheet } from './_fixtures';
+import { createTestsForSpreadsheet } from './_createTestsForSpreadsheet';
+import { extractPropsFromSpreadsheet } from '../../src/extractBlockPropsFromJSONLD/extractPropsFromSpreadsheet';
+
+describe('extractPropsFromSpreadsheet()', () => {
+  createTestsForSpreadsheet(
+    'spreadsheet',
+    spreadsheet,
+    extractPropsFromSpreadsheet,
+  );
+});