+++ packages/media/media-store/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @atlaskit/media-store
 
+## 5.1.0
+- [minor] use context.getFile in media-card [fad25ec](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/fad25ec)
+- [minor] Updated dependencies [fad25ec](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/fad25ec)
+  - @atlaskit/media-test-helpers@15.2.0
+
 ## 5.0.0